Update requirements and fix issues from upgrade (#335)
* Fix taking time slice with localized datetimes

when underlying array is not localized

* fix argument to np.issubdtype in test

* fix pvlib deprecation warnings

* update requirements to be latest

* handle a bunch of deprecation warnings

* update test requirements to latest

* update whatsnew

* freeze requirements

* add comments

+++ solarforecastarbiter/metrics/tests/test_preprocessing.py
@@ -25,7 +25,8 @@ THREE_HOURS_NAN = THREE_HOURS[[True, False, True]]
 THREE_HOURS_EMPTY = pd.DatetimeIndex([], name="timestamp", freq="60min",
                                      tz="MST")
 
-THREE_HOUR_EMPTY_SERIES = pd.Series([], index=THREE_HOURS_EMPTY, name="value")
+THREE_HOUR_EMPTY_SERIES = pd.Series([], index=THREE_HOURS_EMPTY, name="value",
+                                    dtype='float64')
 
 EMPTY_OBJ_SERIES = pd.Series(
     [],
